requirements.txt
Flask
Flask-Cors
mysql-connector-python
bcrypt
google-generativeai